+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 2024-07-24
+- Add erl::common::PangolinPlotterImage
+
 # 2024-07-22
 - Add Tracy
 - Update setup.py to set `ERL_USE_TRACY=OFF` by default